Testing out xml-processing of sheets
So far, this process seems to be working. The only hangups will be the processing of strings, but I should be able to create a template with params and use call-templates on the strings themselves.

This will avoid the ginormous number of duplicated namespaces added to the documents and gives me a path for making things more transferrable.