fix condition in set_ktag_block
There was a challenge block with three solution blocks that was
having an issue where the last solution block was not being acknowledged
as a solution block. My condition caught the block correctly (one block
with two tags), but when it went to find the correct solution block, it
got confused because it was just looking for the first block quote, so I
paramaterized it to be the first block quote that was NOT tagged.

> ## Some challenge
>
> > ## Solution 1
> > ...
> {: .solution}
>
> text ... text
>
> > ## Solution 3
> > ...
> {: .solution}
{: .challenge}