fix duplicate heading display
There was a tricky situation because {cli} did assumed the tree to be
unordered, so any sub-headings of duplicate headings would be masked if
they were lower on the tree than previously duplicated headings.

This introduces a partial fix for this issue, but it will pop up again
if there are multiple duplicate headings with an equal number of
subheadings.