fix internal link testing
This addresses @fmichonneau's comment
https://github.com/carpentries/pegboard/pull/44#discussion_r675644157

In this iteration, I'm now cleaning the URLs and testing if they
actually exist in the current lesson. This also benefits from being a
faster method than querying all the files in a lesson and trying to
manipulate them to fit all the possible iterations of the links.